Piped function calls

+++ src/runtime/tests/runtime_test_utils.rs
@@ -70,11 +70,11 @@ where
 pub fn number_list<T>(values: &[T]) -> Value
 where
     T: Copy,
-    f64: From<T>,
+    i64: From<T>,
 {
     let values = values
         .iter()
-        .map(|n| Number(f64::from(*n).into()))
+        .map(|n| Number(i64::from(*n).into()))
         .collect::<Vec<_>>();
     value_list(&values)
 }
@@ -82,11 +82,11 @@ where
 pub fn number_tuple<T>(values: &[T]) -> Value
 where
     T: Copy,
-    f64: From<T>,
+    i64: From<T>,
 {
     let values = values
         .iter()
-        .map(|n| Number(f64::from(*n).into()))
+        .map(|n| Number(i64::from(*n).into()))
         .collect::<Vec<_>>();
     value_tuple(&values)
 }
